Removed explicit strategy on the GeneratedValue annotation, since it is the default anyway.
Changed the Scala code to use case classes. Also added the toString method that both the Java and Kotlin example have on Pet.

NOTE: For the Hibernate annotations to work on class parameters you have to use a Scala 'field' annotation. It is possible to create type aliases, to make this look nicer, but it feels a bit like cheating.